Additions 4 + Arrange all the classes
added levels into the game and rearrange the classes

+++ Obstacle.java
@@ -3,7 +3,7 @@ package p4_group_8_repo;
 import javafx.scene.image.Image;
 
 public class Obstacle extends Actor {
-	private int speed;
+	private double speed;
 	
 	@Override
 	public void act(long now) {
@@ -14,7 +14,7 @@ public class Obstacle extends Actor {
 			setX(600);
 	}
 	
-	public Obstacle(String imageLink, int xpos, int ypos, int s, int w, int h) {
+	public Obstacle(String imageLink, int xpos, int ypos, double s, int w, int h) {
 		setImageXY(xpos,ypos,new Image(imageLink, w,h, true, true));
 		speed = s;
 	}
