+++ WetTurtle.java
@@ -7,12 +7,11 @@ public class WetTurtle extends Actor{
 	private Image turtle2;
 	private Image turtle3;
 	private Image turtle4;
-	private int speed;
+	private double speed;
 	private boolean sunk = false;
 	
 	@Override
 	public void act(long now) {
-
 		if (now/900000000  % 4 ==0) {
 			setImage(turtle2);
 			sunk = false;
@@ -36,7 +35,7 @@ public class WetTurtle extends Actor{
 			setX(600);
 	}
 	
-	public WetTurtle(int xpos, int ypos, int s, int w, int h) {
+	public WetTurtle(int xpos, int ypos, double s, int w, int h) {
 		turtle1 = new Image("file:src/p4_group_8_repo/resources/TurtleAnimation1.png", w, h, true, true);
 		turtle2 = new Image("file:src/p4_group_8_repo/resources/TurtleAnimation2Wet.png", w, h, true, true);
 		turtle3 = new Image("file:src/p4_group_8_repo/resources/TurtleAnimation3Wet.png", w, h, true, true);
